Migrate ContentTitle.jsx to tsx
